reorganize and add api for rangedSensor
