+++ src/main/java/datagenerator/rangedsensor/api/Cell.java
@@ -1,0 +1,9 @@
+package datagenerator.rangedsensor.api;
+
+public interface Cell {
+    int getCellNumber();
+
+    double getFromTime();
+
+    RangeSensor getLevel();
+}
