+++ src/main/java/datagenerator/rangedsensor/api/RangeSensor.java
@@ -1,0 +1,12 @@
+package datagenerator.rangedsensor.api;
+
+public interface RangeSensor {
+    int getLowerBound();
+
+    int getUpperBound();
+
+    //TODO
+    default byte[] getValue() {
+        return new byte[] {0};
+    }
+}
